tutorials/tutorial_04_application_parameters/python/requirements.txt
paho_mqtt==1.6.1
pyzediot==0.1
